// frontend/src/types/ResponseFlashcard.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ResponseFlashcard = { id: bigint, question: string, answer: string, type: string, is_retired: boolean, is_stale: boolean, easiness_factor: number, interval: number, repetitions: number, next_review: string, last_reviewed: string, };